feat: org/move chat components and global files
