+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { useRef, useEffect, useCallback, useSyncExternalStore } from 'react';
-import { ChatView } from './components/ChatView';
-import { getToken } from './utils/token/token';
+import { ChatView } from './core/components/ChatView';
+import { getToken } from './core/utils/token/token';
 import OpenAI from 'openai';
 import {
   isContentChunk,
