+++ src/features/blockchain/utils/chat/helpers.ts
@@ -1,7 +1,4 @@
-import {
-  CoinRecord,
-  ERC20Record,
-} from '../../features/blockchain/config/chainsRegistry';
+import { CoinRecord, ERC20Record } from '../../config/chainsRegistry';
 //  todo - convert to use interface
 export const getStoredMasks = (): {
   masksToValues: Record<string, string>;
