+++ src/main.tsx
@@ -7,7 +7,7 @@ import { createRoot } from 'react-dom/client';
 //
 // Global CSS & Fonts
 //
-import './global.css';
+import './shared/theme/global.css';
 
 import '@fontsource/inter/300.css';
 import '@fontsource/inter/300-italic.css';
