+++ src/stories/ChatView.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import { fn } from '@storybook/test';
-import { ChatView, ChatViewProps } from '../components/ChatView';
+import { ChatView, ChatViewProps } from '../core/components/ChatView';
 import { mockChatMessages } from './mockdata';
 import { defaultIntroText } from '../features/blockchain/config';
 import type { Decorator } from '@storybook/react';
